ask/ask detail/change - edit-dialog add

+++ Client/src/app/DOM/Ask/ask.component.ts
@@ -5,6 +5,10 @@ import { MatTableDataSource } from '@angular/material/table';
 import { animate, state, style, transition, trigger } from '@angular/animations';
 import { SharedService } from 'src/app/Services/shared.service';
 import { environment } from 'src/environments/environment';
+import { UserDetailsViewComponent } from 'src/app/DOM/Account/user-details-view/user-details-view.component';
+import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
+import { ConfirmDialogComponent } from 'src/app/DOM/Shared/confirm-dialog/confirm-dialog.component';
+import { EditDialogComponent } from '../Shared/edit-dialog/edit-dialog.component';
 
 export interface Article {
   date: any;
@@ -36,6 +40,7 @@ export interface Article {
 export class AskComponent implements AfterViewInit {
   askTitle: string;
   askDescription: string;
+  content: string;
 
   askBoardPkg: any = {
     date: new Date(),
@@ -44,12 +49,27 @@ export class AskComponent implements AfterViewInit {
     userId: '',
   };
 
+  askEditPkg: any = {
+    id: 0,
+    date: new Date(),
+    title: '',
+    description: '',
+    userId: '',
+    parentId: 0,
+  };
+
   active = 1;
   displayedColumns: string[] = ['id', 'userName', 'title', 'date', 'edit'];
   dataSource: MatTableDataSource<Article>;
   expandedElement: Article | null;
 
-  @ViewChild(MatPaginator) paginator: MatPaginator;
+  //@ViewChild(MatPaginator) paginator: MatPaginator;
+  @ViewChild(MatPaginator, { static: false })
+  set paginator(value: MatPaginator) {
+    if (this.dataSource) {
+      this.dataSource.paginator = value;
+    }
+  }
   @ViewChild(MatSort) sort: MatSort;
 
   userId = '';
@@ -57,8 +77,9 @@ export class AskComponent implements AfterViewInit {
   filePath = environment.PhotoFileUrl;
   articles: any = [];
   page = 1;
+  count = 0;
 
-  constructor(private service: SharedService) {
+  constructor(private service: SharedService, public dialog: MatDialog) {
     // Create 100 users
     //const users = Array.from({length: 100}, (_, k) => createNewUser(k + 1));
     // Assign the data to the data source for the table to render;
@@ -68,16 +89,16 @@ export class AskComponent implements AfterViewInit {
     this.userId = this.service.isLoginUser;
     this.userId = this.userId.replace(/['"]+/g, '');
 
+    this.askBoardPkg.title = '';
+    this.askBoardPkg.description = '';
+    this.content = '';
     this.loadUserItem();
   }
 
   loadUserItem() {
     this.service.getArticleList().subscribe((data: any) => {
       this.articles = data;
-      console.log(this.articles);
       this.dataSource = new MatTableDataSource(this.articles);
-      console.log(this.dataSource);
-      console.log(this.dataSource.paginator);
     });
   }
 
@@ -97,15 +118,77 @@ export class AskComponent implements AfterViewInit {
 
   requestAsk() {
     this.askBoardPkg.userId = this.userId;
+    this.askBoardPkg.description = this.content;
     console.log(this.askBoardPkg);
-    // this.service.insertArticle(this.askBoardPkg).subscribe((data: any) => {
-    //   this.ngOnInit();
-    // });
+    this.service.insertArticle(this.askBoardPkg).subscribe((data: any) => {
+      this.ngOnInit();
+    });
   }
 
   onDelete(id: any) {
-    this.service.deleteArticle(id).subscribe((data: any) => {
-      this.ngOnInit();
+    this.count = 0;
+    this.service.getArticleWithReply(id).subscribe((data: any) => {
+      this.count = data.length - 1;
+      if (this.count > 0) this.service.alert('success', 'This asked content cannot be deleted because a reply exists.');
+      else {
+        const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+          width: '350px',
+          data: {
+            title: 'Confirm deletion',
+            message: 'Want to delete the asked content?<br/> Click Yes, The asked content will be deleted',
+          },
+        });
+        dialogRef.afterClosed().subscribe((result) => {
+          if (result) {
+            this.service.deleteArticle(id).subscribe((data: any) => {
+              this.ngOnInit();
+            });
+          }
+        });
+      }
+    });
+  }
+
+  onChange(content: string) {
+    this.content = content;
+  }
+
+  openBorrowerDetails(id: any) {
+    const dialogRef = this.dialog.open(UserDetailsViewComponent, {
+      // height: '500px',
+      width: '300px',
+      data: {
+        dataKey: id,
+      },
+    });
+    dialogRef.afterClosed().subscribe((result) => {
+      console.log(`Dialog result: ${result}`);
+    });
+  }
+
+  onEdit(id: number, content: string, title: string) {
+    const dialogRef = this.dialog.open(EditDialogComponent, {
+      height: '250px',
+      width: '800px',
+      data: {
+        content: content,
+      },
+    });
+
+    dialogRef.afterClosed().subscribe((data: any) => {
+      if (data) {
+        this.askEditPkg.id = id;
+        this.askEditPkg.userId = this.userId;
+        this.askEditPkg.description = data;
+        this.askEditPkg.title = title;
+        this.askEditPkg.parentId = id;
+
+        console.log(this.askEditPkg);
+        this.service.updateArticle(this.askEditPkg).subscribe((data: any) => {
+          this.service.alert('success', 'The content is changed.');
+          this.ngOnInit();
+        });
+      }
     });
   }
 }
